add db dump

+++ src/main/java/ru/aleksanderSil4enko/todoproject/model/User.java
@@ -11,17 +11,14 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    private int id;
+    private long id;
 
     @Column(name = "first_name")
     private String firstName;
-
     @Column(name = "last_name")
     private String lastName;
-
     @Column(name = "email")
     private String email;
-
     @Column(name = "password")
     private String password;
